bugfix
Each cell calculates the power of its associated rays. To do this, the function traverses the cells affected by the rays. This procedure also runs in the initialization phase of the cells. Because initialization occurs on multiple threads, the cell that the calculation would retrieve may not have been created yet, causing the program to crash. The error is fixed by the fix: the initialization step runs in two parts: first the cells are created, and then, when they are all done, the other calculations run.